# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: AgentWire API
  version: 2.0.0
  description: >
    API for AgentWire: AI traffic ingestion, security scanning, replay,
    RBAC/SSO-authenticated access, and audit logging. Single organization per
    self-hosted instance (see docs/architecture.md).
security:
  - BearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      summary: Local email/password login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '401': { description: "Invalid credentials (generic - same response for wrong password or unknown email)." }
  /v1/auth/oidc/login:
    get:
      summary: Redirects to the configured OIDC Authority. 404 if OIDC isn't configured.
      security: []
      responses:
        '302': { description: Redirect to IdP. }
        '404': { description: OIDC not configured on this instance. }
  /v1/auth/oidc/complete:
    get:
      summary: OIDC callback target. Provisions/finds the local user and issues a JWT.
      security: []
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '400': { description: Provisioning failed (e.g. no email claim on the external identity). }
...
  /saml/login:
    get:
      summary: Redirects to the configured SAML IdP. 404 if no IdP metadata is loaded.
      security: []
      responses:
        '302': { description: Redirect to IdP. }
        '404': { description: SAML not configured on this instance. }
  /saml/acs:
    post:
      summary: SAML Assertion Consumer Service. Validates the signed response and issues a JWT.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                SAMLResponse: { type: string }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '400': { description: "Invalid SAML response (signature, status, or provisioning failure) - generic message, no internal detail leaked." }
...
  /v1/audit-log:
    get:
      summary: Append-only audit log. Admin-only. No PUT/PATCH/DELETE route exists under this path.
      parameters:
        - { name: eventType, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditLogEntry' }
        '403': { description: Caller is not an Admin. }
  /v1/users:
    get:
      summary: List users in the caller's organization. Admin-only.
      responses:
        '200': { description: Success. }
    post:
      summary: Create a user with a temporary password. Admin-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, tempPassword, role]
              properties:
                email: { type: string }
                tempPassword: { type: string, minLength: 8 }
                role: { type: string, enum: [Member, Admin] }
      responses:
        '201': { description: Created. }
        '409': { description: Email already in use. }
  /v1/users/{id}/role:
    patch:
      summary: Change a user's role. Admin-only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [Member, Admin] }
      responses:
        '200': { description: Success. }
        '404': { description: User not found. }
  /v1/apikeys:
    get:
      summary: List API keys (prefix + metadata only, never the raw key). Admin-only.
      responses:
        '200': { description: Success. }
    post:
      summary: Create a new API key. Admin-only.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '201':
          description: "Created - the raw key is returned once and never again."
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  key: { type: string }
  /v1/apikeys/{id}/revoke:
    post:
      summary: Revoke an API key. Admin-only.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Success. }
        '404': { description: Key not found. }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-issued JWT (local login, or find-or-provisioned via OIDC/SAML).
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Per-organization ingestion key, used only by POST /v1/traces.
  schemas:
...
    AuditLogEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        eventType: { type: string }
        actorEmail: { type: string, nullable: true }
        targetType: { type: string, nullable: true }
        targetId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }